Password : relocating types to src/types/[typeName]
