+++ packages/password/__tests__/utils/encryption.ts
@@ -1,8 +1,4 @@
-import {
-  bcryptPassword,
-  hashPassword,
-  verifyPassword,
-} from '../src/encryption';
+import { bcryptPassword, hashPassword, verifyPassword } from '../../src/utils/encryption';
 
 describe('encryption', () => {
   describe('bcryptPassword', () => {
