+++ packages/password/src/types/PasswordCreateUserType.ts
@@ -1,0 +1,6 @@
+import { CreateUserType } from '@accounts/common';
+import { PasswordType } from './PasswordType';
+
+export interface PasswordCreateUserType extends CreateUserType {
+  password: PasswordType;
+}
