+++ packages/password/src/types/PasswordLoginType.ts
@@ -1,0 +1,7 @@
+import { LoginUserIdentityType } from '@accounts/common';
+import { PasswordType } from './PasswordType'
+
+export interface PasswordLoginType {
+  user: string | LoginUserIdentityType;
+  password: PasswordType;
+}
