+++ packages/password/src/types/PasswordType.ts
@@ -1,0 +1,8 @@
+import { HashAlgorithm } from '@accounts/common';
+
+export type PasswordType =
+  | string
+  | {
+      digest: string;
+      algorithm: HashAlgorithm;
+    };
